Update type mapping to fix handling of timestamps for Drill 1.19.
Without this patch, queries like SELECT CURRENT_DATE and SELECT
CURRENT_TIMESTAMP return the wrong values when used with Drill 1.19.
Presumably this is due to a change in the way the dates are represented
in the REST API.  The wrong values result from an incorrect assumption
of nanosecond precision in UNIX timestamps when Drill in fact returns
millisecond precision UNIX timestamps.